api/serializers.py
from rest_framework import serializers
from .models import Producto


class ProductoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Producto
        fields = ['id', 'nombre', 'descripcion', 'precio'] # como un DTO de Spring en Java, filtra que mostrar en la respuesta

